Move serve.ts to top level
