+++ http/mod.ts
@@ -1,2 +1,1 @@
 export * from "./handler.ts";
-export * from "./server.ts";
